+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# 2.0.5 - 2023-12-15
+
+## Added
+
+### [#1336] Tweaks for sdk metrics
+Shielded verified and total balances are logged for every sync of `SDKMetrics`. 
+
+## Checkpoints
+
+Mainnet
+Sources/ZcashLightClientKit/Resources/checkpoints/mainnet/2330000.json
+
+Testnet
+Sources/ZcashLightClientKit/Resources/checkpoints/testnet/2630000.json
+
 # 2.0.4 - 2023-12-12
 
 ## Changed